bookmybook/requirements.txt
Django
gunicorn
djangorestframework
mysqlclient
django-mysql
redis
kafka-python